Publish: Use package name minus "spl-" instead of dir for tag (#55)
Publish: Use package name minus "spl-" instead of dir

#### Problem

The repo has some nested packages, such as `program-error/derive`, but
the publish script will use the basename for the tag name. This means
that we'll publish a tag called `derive`, rather than
`program-error-derive`.

#### Summary of changes

Change the tag name to the package name minus `spl-`. Also, the publish
commit is a generic message, so customize it.